adk_lab_05/illustrator_agent/requirements.txt
google-adk==1.21.0
a2a-sdk==0.3.21